Move sx tests to sx package. closes #1672
