+++ silx/sx/test/__init__.py
@@ -1,7 +1,6 @@
 # coding: utf-8
 # /*##########################################################################
-#
-# Copyright (c) 2015-2018 European Synchrotron Radiation Facility
+# Copyright (C) 2018 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -21,34 +20,19 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #
-# ###########################################################################*/
+# ############################################################################*/
 
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "30/03/2017"
+__date__ = "06/03/2018"
 
-from numpy.distutils.misc_util import Configuration
-
-
-def configuration(parent_package='', top_path=None):
-    config = Configuration('silx', parent_package, top_path)
-    config.add_subpackage('gui')
-    config.add_subpackage('io')
-    config.add_subpackage('math')
-    config.add_subpackage('image')
-    config.add_subpackage('opencl')
-    config.add_subpackage('resources')
-    config.add_subpackage('sx')
-    config.add_subpackage('sx.test')
-    config.add_subpackage('test')
-    config.add_subpackage('third_party')
-    config.add_subpackage('utils')
-    config.add_subpackage('app')
-
-    return config
+import unittest
 
 
-if __name__ == "__main__":
-    from numpy.distutils.core import setup
+from . import test_sx
 
-    setup(configuration=configuration)
+
+def suite():
+    test_suite = unittest.TestSuite()
+    test_suite.addTest(test_sx.suite())
+    return test_suite
