+++ silx/sx/test/test_sx.py
